updated code changes for UNIRUB

+++ TYR_2.0_18.5/TY_69_START_UNIRUB_ADHESION.py
@@ -2344,7 +2344,7 @@ class PlotCanvas_Auto(FigureCanvas):
         results=connection.execute("SELECT GRAPH_SCALE_CELL_2,GRAPH_SCALE_CELL_1,AUTO_REV_TIME_OFF,BREAKING_SENCE,ISACTIVE_MODBUS,MODBUS_PORT,NON_MODBUS_PORT from SETTING_MST") 
         for x in results:
                  self.auto_rev_time_off=int(x[2])
-                 self.break_sence=float(x[3])
+                 self.break_sence=int(x[3])
                  self.modbus_flag=str(x[4])
                  self.modbus_port=str(x[5])
                  self.non_modbus_port=str(x[6])
@@ -2871,8 +2871,8 @@ class PlotCanvas_Auto(FigureCanvas):
                  self.calc_speed=(int(self.input_speed_val)/int(self.speed_val))*1000                 
                  #print(" calc Speed : "+str(self.calc_speed))
                  #print(" command: *P"+str(self.calc_speed)+" \r")
-                 #self.command_str="*P%04d"%self.calc_speed+"_%04d"%self.break_sence+"\r"
-                 self.command_str="*P%04d"%self.calc_speed+"_%0.2f"%self.break_sence+"\r"
+                 self.command_str="*P%04d"%self.calc_speed+"_%04d"%self.break_sence+"\r"
+                 #self.command_str="*P%04d"%self.calc_speed+"_%0.2f"%self.break_sence+"\r"
                  print("Morot Speed and Breaking speed Command  :"+str(self.command_str))
             else:
                  print(" not Ok ")
